#71 feat: 할일 추가 새로고침 반영, 옵션 개별 클릭 기능
