+++ src/layout/Layout.jsx
@@ -44,4 +44,5 @@ const PageContainer = styled.div`
   height: 85vh;
   margin-top: 17%;
   padding: 0 2% 4.05vh 2%;
+  overflow: auto;
 `;
