+++ src/shared/GlobalStyle.jsx
@@ -29,7 +29,9 @@ body, div, button{
   max-width: 412px;
   max-height: 915px;
 }
-
+input {
+  padding-left: 1em;
+}
 button {
   cursor: pointer;
   outline: none;
